DroneGCS/pyproject.toml
[project]
name = "drone-gcs"
version = "1.0.0"
description = "Ground Control Station for Drone Firmware"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.5.0",
    "pymavlink>=2.4.0",
    "pyserial>=3.5",
    "numpy>=1.24.0",
    "pyqtgraph>=0.13.0",
    "pygame-ce>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["drone_gcs"]
